Reduce load by queueing and simple rate-limiting.
Instead of directly sending outgoing mqtt messages, put them
in an ordered dictionary that is treated like a FIFO. Send
messages from the FIFO using an idle event handler. If another
value for the same topic is received from dbus while the previous
one is still in the queue, overwrite it. This implements a simple
form of rate-limiting.

Additionally, when the FIFO is empty, pause for 2 seconds. This
also increases the chance that a more recent value for the same topic
might be sent, reducing the volume at the slight cost of a delay, improving the
rate limiting.

Also includes other mild improvements such as wrapping the main procedural code
in a `def main`.

victronenergy/venus#189