stop writing indefinetly to a broken socket
If the connection to the broker errors, e.g. since the broker got restarted,
the process can end up in an infinite loop when trying to send data to a non
existing connection. So add the missing error handling to prevent this.

svc -t /service/gui can trigger this for example.