ignore the payload on system/0/Serial
A payload was never expected on /Serial, so initially the ability
to send a payload was also implemented here. However, it turns out
some applications in the field send a payload that doesn't parse
properly and this causes a regression for them.

So now we ignore the payload on /Serial. People who want to use the
new model must use /keepalive.